[강운덕] [PINPOINT-267] histogram의 timeslot 값이 int와 short이 혼용되는 부분이 있어 int로 수정함.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-commons/trunk@3235 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/common/HistogramSlot.java
@@ -4,11 +4,11 @@ package com.nhn.pinpoint.common;
  * @author emeroad
  */
 public class HistogramSlot {
-    private final int slotTime;
+    private final short slotTime;
     private final ResponseCode responseCode;
 
 
-    public HistogramSlot(int slotTime, ResponseCode responseCode) {
+    public HistogramSlot(short slotTime, ResponseCode responseCode) {
         if (responseCode == null) {
             throw new NullPointerException("responseCode must not be null");
         }
@@ -17,7 +17,7 @@ public class HistogramSlot {
     }
 
     // TODO : int를 short으로 바꿔야할 듯...
-    public int getSlotTime() {
+    public short getSlotTime() {
         return slotTime;
     }
